requirements.txt
flask
torch
diffusers
Pillow
